chore(refactor) refactoring the logic of the backend [FINISHES #167216160]
