+++ src/middleware/auth.js
@@ -1,5 +1,4 @@
 import jwt from 'jsonwebtoken';
-import config from 'config';
 
 
 export default (req, res, next) => {
@@ -7,7 +6,7 @@ export default (req, res, next) => {
   if (!token) return res.status(401).send('Access denied, No token provided');
 
   try {
-    const decoded = jwt.verify(token, config.get('jwtPrivateKey'));
+    const decoded = jwt.verify(token, process.env.JWT_PRIVATE_KEY);
     req.user = decoded;
     next();
   } catch (ex) {
