+++ src/test/routes/signin.test.js
@@ -1,9 +1,8 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
 import jwt from 'jsonwebtoken';
-import config from 'config';
 import bcrypt from 'bcrypt';
-import { users } from '../../models/user';
+import { users } from '../../helper/user';
 import server from '../../index';
 
 chai.use(chaiHttp);
@@ -56,7 +55,7 @@ describe('auth/signin', () => {
 
       const res = await exec();
 
-      const decoded = jwt.verify(res.body.data.token, config.get('jwtPrivateKey'));
+      const decoded = jwt.verify(res.body.data.token, process.env.JWT_PRIVATE_KEY);
 
       expect(res.status).to.equal(200);
       expect(decoded).to.have.property('isAdmin');
