+++ src/test/routes/signup.test.js
@@ -1,8 +1,7 @@
 import chaiHttp from 'chai-http';
 import chai from 'chai';
 import jwt from 'jsonwebtoken';
-import config from 'config';
-import { users } from '../../models/user';
+import { users } from '../../helper/user';
 import server from '../../index';
 
 chai.use(chaiHttp);
@@ -50,7 +49,7 @@ describe('auth/signup', () => {
     it('should return 200 if user is registered successfully', async () => {
       const res = await exec();
 
-      const decoded = jwt.verify(res.body.data.token, config.get('jwtPrivateKey'));
+      const decoded = jwt.verify(res.body.data.token, process.env.JWT_PRIVATE_KEY);
 
       expect(res.status).to.equal(200);
       expect(decoded).to.have.property('isAdmin');
